+++ types/Fill.h
@@ -9,7 +9,7 @@
 
 struct Fill : public Event
 {
-    Fill(OrderId order_id, Type type, Side side, uint32_t instrument_id, float price, float quantity, float fee,
+    Fill(OrderId order_id, Type type, Side side, uint32_t instrument_id, double price, double quantity, double fee,
          uint64_t timestamp)
         : Event(FILL, timestamp), order_id_(order_id), type_(type), side_(side), instrument_id_(instrument_id),
           price_(price), quantity_(quantity), fee_(fee), timestamp_(timestamp)
@@ -20,9 +20,9 @@ struct Fill : public Event
     Type type_;
     Side side_;
     uint32_t instrument_id_;
-    float price_;
-    float quantity_;
-    float fee_;
+    double price_;
+    double quantity_;
+    double fee_;
     uint64_t timestamp_;
 };
