+++ types/Order.h
@@ -20,7 +20,8 @@ enum class Side
 
 struct Order : public Event
 {
-    Order(OrderId orderId, Type type, Side side, int32_t instrument_id, float price, float quantity, uint64_t timestamp)
+    Order(OrderId orderId, Type type, Side side, int32_t instrument_id, double price, double quantity,
+          uint64_t timestamp)
         : Event(ORDER, timestamp), orderId_(orderId), type_(type), side_(side), instrument_id_(instrument_id),
           price_(price), quantity_(quantity), timestamp_(timestamp)
     {
@@ -30,8 +31,8 @@ struct Order : public Event
     Type type_;
     Side side_;
     int32_t instrument_id_;
-    float price_;
-    float quantity_;
+    double price_;
+    double quantity_;
     uint64_t timestamp_;
 };
