+++ types/Position.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct Position {
-  float avg_price{0};
-  float realized_pnl{0};
-  float quantity{0};
+  double avg_price{0.0f};
+  double realized_pnl{0.0f};
+  double quantity{0.0f};
 };
