+++ types/Signal.h
@@ -7,7 +7,7 @@
 
 struct Signal : public Event
 {
-    Signal(Type type, Side side, int32_t instrument_id, float price, float quantity, uint64_t timestamp)
+    Signal(Type type, Side side, int32_t instrument_id, double price, double quantity, uint64_t timestamp)
         : Event(SIGNAL, timestamp), type_(type), side_(side), instrument_id_(instrument_id), price_(price),
           quantity_(quantity), timestamp_(timestamp)
     {
@@ -16,8 +16,8 @@ struct Signal : public Event
     Type type_;
     Side side_;
     int32_t instrument_id_;
-    float price_;
-    float quantity_;
+    double price_;
+    double quantity_;
     uint64_t timestamp_;
 };
